src/main/java/Shop.java
import accessories.ISell;

import java.util.ArrayList;

public class Shop {

    private ArrayList<ISell> stock;

    public Shop(ArrayList<ISell> stock) {
        this.stock = new ArrayList<ISell>();
    }

    public ArrayList<ISell> getStock() {
        return stock;
    }


    public void setStock(ArrayList<ISell> stock) {
        this.stock = stock;
    }

    public void addToStock(ISell item){
            this.stock.add(item);
    }

    public void removeFromStock(ISell item){
        this.stock.remove(stock.indexOf(item));
    }

}
